+++ codex/librarian/covers/create.py
@@ -19,9 +19,9 @@ class CoverCreateMixin(CoverPathMixin):
     """Create methods for covers."""
 
     _COVER_RATIO = 1.5372233400402415  # modal cover ratio
-    _THUMBNAIL_WIDTH = 165
-    _THUMBNAIL_HEIGHT = round(_THUMBNAIL_WIDTH * _COVER_RATIO)
-    _THUMBNAIL_SIZE = (_THUMBNAIL_WIDTH, _THUMBNAIL_HEIGHT)
+    THUMBNAIL_WIDTH = 165
+    THUMBNAIL_HEIGHT = round(THUMBNAIL_WIDTH * _COVER_RATIO)
+    _THUMBNAIL_SIZE = (THUMBNAIL_WIDTH, THUMBNAIL_HEIGHT)
 
     @classmethod
     def _create_cover_thumbnail(cls, cover_image_data):
@@ -81,19 +81,8 @@ class CoverCreateMixin(CoverPathMixin):
         else:
             cover_path.symlink_to(self.MISSING_COVER_PATH)
 
-    #####################
-    # UNUSED BELOW HERE #
-    #####################
-
-    def create_cover(self, pk):
-        """Create a cover from a comic id."""
-        # XXX Unused.
-        cover_path = self.get_cover_path(pk)
-        self.create_cover_from_path(pk, cover_path, self.log, self.librarian_queue)
-
-    def bulk_create_comic_covers(self, comic_pks):
+    def _bulk_create_comic_covers(self, comic_pks):
         """Create bulk comic covers."""
-        # XXX Unused
         num_comics = len(comic_pks)
         if not num_comics:
             return None
@@ -111,7 +100,9 @@ class CoverCreateMixin(CoverPathMixin):
                     status.decrement_total()
                 else:
                     # bulk creator creates covers inline
-                    self.create_cover(pk)
+                    self.create_cover_from_path(
+                        pk, cover_path, self.log, self.librarian_queue
+                    )
                     status.increment_complete()
                 self.status_controller.update(status)
 
@@ -120,3 +111,8 @@ class CoverCreateMixin(CoverPathMixin):
         finally:
             self.status_controller.finish(status)
         return status.complete
+
+    def create_all_covers(self):
+        """Create all covers for all libraries."""
+        pks = Comic.objects.values_list("pk", flat=True)
+        self._bulk_create_comic_covers(pks)