+++ codex/views/opds/__init__.py
@@ -1,0 +1,1 @@
+"""OPDS Views Common to all versions."""